daily_crypto_news/daily_market_report/pyproject.toml
[project]
name = "daily_market_report"
version = "0.1.0"
description = "daily_market_report using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.14.1",
    "yfinance>=0.2.40",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "jinja2>=3.1.0",
    "weasyprint>=62.0",
    "playwright>=1.40.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "feedparser>=6.0.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.8.0",
]

[project.scripts]
kickoff = "daily_market_report.main:kickoff"
run_crew = "daily_market_report.main:kickoff"
plot = "daily_market_report.main:plot"
run_with_trigger = "daily_market_report.main:run_with_trigger"
daily-market-report = "daily_market_report.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
